airflow_docker_1/requirements.txt
clickhouse-connect==0.7.8
clickhouse-driver==0.2.7
airflow-providers-clickhouse==0.0.1